+++ src/main/java/appeng/items/tools/powered/ToolPortableCell.java
@@ -118,8 +118,8 @@ public class ToolPortableCell extends AEBasePoweredItem implements IBasicCellIte
     }
 
     @Override
-    public AEKeyType getKeyType() {
-        return AEKeyType.items();
+    public Set<AEKeyType> getKeyTypes() {
+        return Collections.singleton(AEKeyType.items());
     }
 
     @Override
@@ -139,7 +139,7 @@ public class ToolPortableCell extends AEBasePoweredItem implements IBasicCellIte
 
     @Override
     public IItemHandler getConfigInventory(final ItemStack is) {
-        return new CellConfig(is, Collections.singleton(this.getKeyType()));
+        return new CellConfig(is, this.getKeyTypes());
     }
 
     @Override